Factor file_utils into its own file, address comments
